Move Serialization object to v1 package
