+++ src/commonTest/kotlin/fr/acinq/lightning/serialization/StateSerializationTestsCommon.kt
@@ -3,6 +3,7 @@ package fr.acinq.lightning.serialization
 import fr.acinq.bitcoin.Block
 import fr.acinq.lightning.Lightning.randomKey
 import fr.acinq.lightning.channel.TestsHelper
+import fr.acinq.lightning.serialization.v1.Serialization
 import fr.acinq.lightning.tests.TestConstants
 import fr.acinq.lightning.tests.utils.LightningTestSuite
 import fr.acinq.secp256k1.Hex
@@ -56,7 +57,7 @@ class StateSerializationTestsCommon : LightningTestSuite() {
         val bin = Hex.decode(
             "bf6776657273696f6e016464617461792c0839663738326136363732326536313633363936653731326536633639363736383734366536393665363732653733363537323639363136633639376136313734363936663665326537363331326534653666373236643631366362663663373337343631373436393633353036313732363136643733626636393633363836313639366534383631373336383738343033303336333233323336363533343336333133313331363133303632333533393633363136313636333133323336333033343333363536323335363236323636333233383633333333343636333336313335363533333333333236313331363636333337363233323632333733333633363633313338333833393331333036363663373236353664366637343635346536663634363534393634373834323330333336363335333633393633363136363336333833313331333633303332363333333632363333363333363336313634333633333632333436343335363336363635333833303338363133333636333233363339363436313333333533343634363236333334333733383332333936363338333536313332333536313633333233326666366136333735373237323635366537343534363937306266363536363639373237333734313930323937363637333635363336663665363437393030613033303330333033303330333033323330363633323333333036323334363336323333363533343339333833313339333533393330333133313631333836323337333133383632363333393330363333373631333636313330333733333334333633323335363633343635333036343337333036363632363636343336333433373631333533313635363533303334333633353331333833353338333533343334333333353632363236343333363433323339333436363636333433363332333533393631333533363635363133353335333736333332363636343332333436363633333036333632333233383636333633393333333133363337333933343332333633363334333536353335333836333334333636353631333836313336333036363636363636363337363633323330333033313330333033303330333033306666373636333735373237323635366537343466366534333638363136393665343636353635373236313734363537336266373236643735373437353631366334333663366637333635343636353635373236313734363562663637363636353635373236313734363531613030303039323763666637303633366336313639366434643631363936653436363536353732363137343635626636373636363536353732363137343635316130303030393237636666366236363631373337343436363536353732363137343635626636373636363536353732363137343635316130303030393237636666666636623633366636643664363937343664363536653734373362663665363336383631366536653635366335363635373237333639366636656266363436323639373437333638333033303330333033303330333036366666366236633666363336313663353036313732363136643733626636363665366636343635343936343738343233303333363536353334333433343636333936323338363533323634363533383337333533393631333533303331363333383339363233383335333333323632333636323635333236343633363333353634333836333338363133313332333533323335333636333335333033383636333333313333333133303336333133323331333636653636373536653634363936653637346236353739353036313734363862663634373036313734363839663162313132613465626439343832656263633162353733313236333535626238306564643362326165333539303663303062313039663162356134373665653533353233363432616666666636393634373537333734346336393664363937343139303232323738313836643631373834383734366336333536363136633735363534393665343636633639363736383734346437333631373431623030303030303031326130356632303036653633363836313665366536353663353236353733363537323736363531393032353836623638373436633633346436393665363936643735366462663634366437333631373430316666366237343666353336353663363634343635366336313739626636613735366536343635373236633739363936653637313930376530666637303664363137383431363336333635373037343635363434383734366336333733313831653638363937333436373536653634363537326634373831383634363536363631373536633734343636393665363136633533363337323639373037343530373536323462363537393738326333303330333133343634363636323634333033343331333333383334363136353336333836313631363136333635363236313632363436333634363633343631333933343334333736313334333836353635333733353334363836363635363137343735373236353733626636393631363337343639373636313734363536346266396637383330363637323265363136333639366537313265366336393637363837343665363936653637326534363635363137343735373236353265346637303734363936663665343436313734363134633666373337333530373236663734363536333734626666666666363934643631366536343631373436663732373939663738326536363732326536313633363936653731326536633639363736383734366536393665363732653436363536313734373537323635326535363631373236393631363236633635346336353665363737343638346636653639366636656266666666663638346637303734363936663665363136633966373832383636373232653631363336393665373132653663363936373638373436653639366536373265343636353631373437353732363532653530363137393664363536653734353336353633373236353734626666666666363834663730373436393666366536313663396637383330363637323265363136333639366537313265366336393637363837343665363936653637326534363635363137343735373236353265343236313733363936333464373536633734363935303631373237343530363137393664363536653734626666666666363834663730373436393666366536313663396637383230363637323265363136333639366537313265366336393637363837343665363936653637326534363635363137343735373236353265353737353664363236666266666666663638346637303734363936663665363136633966373832613636373232653631363336393665373132653663363936373638373436653639366536373265343636353631373437353732363532653533373436313734363936333532363536643666373436353462363537396266666666663638346637303734363936663665363136633966373832383636373232653631363336393665373132653663363936373638373436653639366536373265343636353631373437353732363532653431366536333638366637323466373537343730373537343733626666666666363834663730373436393666366536313663396637383263363637323265363136333639366537313265366336393637363837343665363936653637326534363635363137343735373236353265353437323631366437303666366336393665363535303631373936643635366537346266666666663638346637303734363936663665363136636666666666663663373236353664366637343635353036313732363136643733626636363665366636343635343936343738343233303333363633353336333936333631363633363338333133313336333033323633333336323633333633333633363136343336333336323334363433353633363636353338333033383631333336363332333633393634363133333335333436343632363333343337333833323339363633383335363133323335363136333332333236393634373537333734346336393664363937343139303232323738313836643631373834383734366336333536363136633735363534393665343636633639363736383734346437333631373431623030303030303031326130356632303036653633363836313665366536353663353236353733363537323736363530303662363837343663363334643639366536393664373536646266363436643733363137343031666636623734366635333635366336363434363536633631373962663661373536653634363537323663373936393665363731393032643066663730366436313738343136333633363537303734363536343438373436633633373331383165366436363735366536343639366536373530373536323462363537393738343233303333333533373338333133353334333036333335333236313636333936313636363433323634333933393338333033363336333336323338333336363631363533303633363333353339333536343632363133363335363536343632363636333633363136333335333036363631333836313332333736343339333533303339333537333732363537363666363336313734363936663665343236313733363537303666363936653734373834323330333233383339333633373333363133373332333933323339333433393632333036323335363133333633333036363338333333333631333933373635333833303636333233393335333436333636363433323635333036363338333433333330363333303632363433343634333736363338363433323635363536343634333233373730373036313739366436353665373434323631373336353730366636393665373437383432333033323634333236323632333133363332333533323332333933393334333536313633363633343335333036313631333833363635363333393338363436363338363633363634363133343337333933313634333736323333333436343633333933383333333236353338333636343333333533393634333033393337333333353332373736343635366336313739363536343530363137393664363536653734343236313733363537303666363936653734373834323330333333313338333536353632333633353635333336313634363433393631333136333330363436323632363133333635333933303330363236363330333633313634333633393337333936323332333436313336363533303635363633353634333333383635363633383634363333383330333233393336333333333636333133333664363837343663363334323631373336353730366636393665373437383432333033323635333933313335363136333633363636353634333933333330333133333335333236333635333533323337363233393330333333353337363433313334333133363337333233303632333233343631333933303337333933383332333833313334363233363332333036363633363633343332333333343631333536313635363836363635363137343735373236353733626636393631363337343639373636313734363536346266396637383330363637323265363136333639366537313265366336393637363837343665363936653637326534363635363137343735373236353265346637303734363936663665343436313734363134633666373337333530373236663734363536333734626666666666363834663730373436393666366536313663396637383265363637323265363136333639366537313265366336393637363837343665363936653637326534363635363137343735373236353265343336383631366536653635366335323631366536373635353137353635373236393635373362666666666636383466373037343639366636653631366339663738326536363732326536313633363936653731326536633639363736383734366536393665363732653436363536313734373537323635326535363631373236393631363236633635346336353665363737343638346636653639366636656266666666663638346637303734363936663665363136633966373833363636373232653631363336393665373132653663363936373638373436653639366536373265343636353631373437353732363532653433363836313665366536353663353236313665363736353531373536353732363936353733343537383734363536653634363536346266666666663638346637303734363936663665363136633966373832613636373232653631363336393665373132653663363936373638373436653639366536373265343636353631373437353732363532653533373436313734363936333532363536643666373436353462363537396266666666663638346637303734363936663665363136633966373832383636373232653631363336393665373132653663363936373638373436653639366536373265343636353631373437353732363532653530363137393664363536653734353336353633373236353734626666666666363834663730373436393666366536313663396637383330363637323265363136333639366537313265366336393637363837343665363936653637326534363635363137343735373236353265343236313733363936333464373536633734363935303631373237343530363137393664363536653734626666666666363834663730373436393666366536313663396637383230363637323265363136333639366537313265366336393637363837343665363936653637326534363635363137343735373236353265353737353664363236666266666666663638346637303734363936663665363136636666666666663663363336383631366536653635366334363663363136373733303036623663366636333631366334333666366436643639373462663635363936653634363537383032363437333730363536336266363536383734366336333733396666663637363636353635373236313734363562663637363636353635373236313734363531393039633466663637373436663463366636333631366362663634366437333631373431613038653139333238666636383734366635323635366436663734363562663634366437333631373431613031393830313338666666663665373037353632366336393733363836313632366336353534373837336266363836333666366436643639373435343738626636353639366537303735373462663638366637353734353036663639366537343738343833373636333733383338333636353339333836323634363433313336363436323338363233303339363136323335363436333339333233383334333633363335333233383636363136353336333433313339333833313332333133373334363636343332333936343632363433363337363633313631363136353332333933303330333133303330333033303330333036353734373834663735373437383536333736333631363533303332333033303330333033303330333033303330333033323332333033303332333033393333363533363333333133343631363233373338363533373634363433363634333136353331363233373338333733343635333633333336363633313634333336313633333336333635363536333634333233373336333936313338363533333635333233363336363333303632333233373334333133353635333036353663373236353634363536353664353336333732363937303734373930303865333533323332333133303333333533373338333133353334333036333335333236313636333936313636363433323634333933393338333033363336333336323338333336363631363533303633363333353339333536343632363133363335363536343632363636333633363136333335333036363631333836313332333736343339333533303339333533323331333033333632333833323337363536313335333733303330333933383331333736353331333536343331333933373339333433353334363633363633333033373631363533363337363333373632333933303330363333393633333033353338363633383333333636313633333333323330333533393333333733393631333536363631333533323631363566663632373437383739303337383330333233303330333033303330333033303330333033313330333133373636333733383338333636353339333836323634363433313336363436323338363233303339363136323335363436333339333233383334333633363335333233383636363136353336333433313339333833313332333133373334363636343332333936343632363433363337363633313631363136353332333933303330333133303330333033303330333033303330333033393339333833343633333833303330333433343631333033313330333033303330333033303330333033303330333033303332333233303330333233303334363133383338363533333634333833353336333733353337333033333632363133303635333336363636363133393633333636333633333033353631333933323635333336313636333933353634333033303337363133323632363133343331363436323632363433363636363233393332333233383633363136343632333436313330333133303330333033303330333033303330333033303330333033323332333033303332333033383339363136353635333333363332363633383635333436333636333336353636333436343633333233363631333133363338333833333330333936333331363633353337333633313333333933323332363536353337333933313334363433353334333633363334333433333636333336353633333236363333333233363635333533353631333033303330333033303330333033303330333033303330333233323330333033323330333733303636333036323333363436323634333733383335333533303338333836363332363533383330333933343335333636333634363333303336333533373333363333383339333236343631333933353336333533393339333033373334333533313339333336323331333136323633333136363330333936323635333333303339333433363330333233303330333033303330333033303330333033303332333233303330333233303330363136313633333433353636333433393337333833373334333433323634363233363339363436313337363436343636363336333632333633363635333133363635333633343337333536333632333233363335363136323635333533333336333236363333333733363339363233353632363436353333333333383636333033343330333033343338333333303334333533303332333233313330333033383335363536353632333836313334363133373635363336353334363536343631363433373336333033343337333036313636333636313339333133383635363233383631363533343338333733383632333833363335333833363635333736343338363636363337333433313334333936333634363236353335363633303330333233323330333736363339333736333335363436353337363136323330363133353635333533303330333836343337363336323330333933333330333936353331333233343334333336323635333433363334333933323633363136353333333233323335333336363635333833373330363236343337333536313634363433373334333033303331333433373333333033343334333033323332333033313334333936313331363133363632333336323333363636333633333633323332363636363332363633333332333536343636333536323338333533343336363533363633333236323335363533363336333236333338333436343336363533353335333133333338333233393634363433313634333333363337363133383330333233323330333233363632363336323330333636313632333636333632333236353331333033363634363636333333333333393636333836363336333133333335333136323632333236313635333833393332363133323333363433323332333033383337333836313339333333343333363436353332333633343331333136323338363233303331333433373335333233323331333033333335333733383331333533343330363333353332363136363339363136363634333236343339333933383330333633363333363233383333363636313635333036333633333533393335363436323631333633353635363436323636363336333631363333353330363636313338363133323337363433393335333033393335333233313330333336323338333233373635363133353337333033303339333833313337363533313335363433313339333733393334333533343636333636333330333736313635333633373633333736323339333033303633333936333330333533383636333833333336363136333333333233303335333933333337333936313335363636313335333236313635333133383634333433343636333233306666366536383734366336333534373837333431366536343533363936373733396666666666666636633732363536643666373436353433366636643664363937346266363536393665363436353738303236343733373036353633626636353638373436633633373339666666363736363635363537323631373436356266363736363635363537323631373436353139303963346666363737343666346336663633363136636266363436643733363137343161303139383031333866663638373436663532363536643666373436356266363436643733363137343161303865313933323866666666363437343738363936343738343036323636333436333331333233343632333933333330333133323336333236353632363536353339363533373335333836363339333133323333333436333335333533363337333736313336363633353636333933313331333636313331333733343332333033373631333933313633363533343634363636313633333833373738313837323635366436663734363535303635373234333666366436643639373436643635366537343530366636393665373437383432333033333337333136363339333933323633333236333332333033333635363433313333333736333633333736353635333536363335333336313636333333353634363536343338363433343334333433323336333333323337333936333632333533313334333733393339363533343337333433393337333533333636363336363330666636633663366636333631366334333638363136653637363537336266363837303732366637303666373336353634396666663636373336393637366536353634396666663635363136333662363536343966666666663664373236353664366637343635343336383631366536373635373362663638373037323666373036663733363536343966666636353631363336623635363439666666363637333639363736653635363439666666666636663663366636333631366334653635373837343438373436633633343936343030373037323635366436663734363534653635373837343438373436633633343936343031363837303631373936643635366537343733626666663734373236353664366637343635346536353738373434333666366436643639373434393665363636666266363537323639363736383734373834323330333233383633363433343633333033363339363536363334333133393336363136333331333033313631333833313331333833363631363433303635363333323333363133343339333533383636333336333334363336343332333433373338363633343632333533343635363236313333333833373339363333333336363636326666366236333666366436643639373434393665373037353734626636383666373537343530366636393665373437383438333736363337333833383336363533393338363236343634333133363634363233383632333033393631363233353634363333393332333833343336333633353332333836363631363533363334333133393338333133323331333733343636363433323339363436323634333633373636333136313631363533323339333033303331333033303330333033303330363537343738346637353734373835363337363336313635333033323330333033303330333033303330333033303330333233323330333033323330333933333635333633333331333436313632333733383635333736343634333636343331363533313632333733383337333436353336333333363636333136343333363136333333363336353635363336343332333733363339363133383635333336353332333633363633333036323332333733343331333536353330363536633732363536343635363536643533363337323639373037343739303038653335333233323331333033333335333733383331333533343330363333353332363136363339363136363634333236343339333933383330333633363333363233383333363636313635333036333633333533393335363436323631333633353635363436323636363336333631363333353330363636313338363133323337363433393335333033393335333233313330333336323338333233373635363133353337333033303339333833313337363533313335363433313339333733393334333533343636333636333330333736313635333633373633333736323339333033303633333936333330333533383636333833333336363136333333333233303335333933333337333936313335363636313335333236313635666637383161373236353664366637343635353036353732343336663664366436393734366436353665373435333635363337323635373437336266366236623665366637373665343836313733363836353733626637383366333033303330333033303330333033303330333033303330333033303330333033313331333133313331333133313331333133313331333133313331333133313331333133313331333133313331333133313331333133313331333133313331333133313331333133313331333133313331333133313331333133313331373834303633363533343633333633383632363436363337363536363330333336313636363133313634333333343335363636333333333433323330333733393334333133313336363633363335363633313338333833363335363333373337333633343334333033363333363136313336363633313632333933303331363133353331666636393663363137333734343936653634363537383162303030306666666666666666666666656666363936333638363136653665363536633439363437383430333736363337333833383336363533393338363236343634333133363634363233383632333033393631363233353634363333393332333833343336333633353332333836363631363533363334333133393338333133323331333733343636363433323339363436323634333633373636333136313631363533323339333166663665373336383666373237343433363836313665366536353663343936346266363236393634316230303030303031643535306630303031666636363632373537323639363536346634373336333638363136653665363536633431366536653666373536653633363536643635366537346636366436333638363136653665363536633535373036343631373436356266363937333639363736653631373437353732363537393030383033383331333536313635333536363633333133313337363136313331333033333332363436313634363536363339333333393332363636313338363433333335333833393336333533313332333236343332333833383339333233333335333936353336363133333633333933333337333133373635333236353330363133313333363636333336363333333337333833363338333036323334333233323334363636353634333333353334363133393334333433313337333933373331363233313334333833303334333236333332333936333331333533313336333533373331333133333335333633373334333633393332333136343334333833303333363936333638363136393665343836313733363837383430333033363332333233363635333433363331333133313631333036323335333936333631363136363331333233363330333433333635363233353632363236363332333836333333333436363333363133353635333333333332363133313636363333373632333236323337333336333636333133383338333933313330363636653733363836663732373434333638363136653665363536633439363462663632363936343162303030303030316435353066303030316666373037343639366436353733373436313664373035333635363336663665363437333161363038616561353636633664363537333733363136373635343636633631363737333031366336333638363136653665363536633436366336313637373330303666363336633734373634353738373036393732373934343635366337343631626636613735366536343635373236633739363936653637313930303930666636663638373436633633346436393665363936643735366434643733363137346266363436643733363137343031666636623636363536353432363137333635346437333631373462663634366437333631373431393033653866663738313936363635363535303732366637303666373237343639366636653631366334643639366336633639366636653734363837333138363436663638373436633633346436313738363936643735366434643733363137346266363436643733363137343161306137393934363066666666373337323635366436663734363534333638363136653665363536633535373036343631373436356636366436633666363336313663353336383735373436343666373736656636366537323635366436663734363535333638373537343634366637373665663666666666ff"
         )
-        val state = fr.acinq.lightning.serialization.Serialization.deserialize(bin, TestConstants.Alice.nodeParams)
+        val state = Serialization.deserialize(bin, TestConstants.Alice.nodeParams)
         println(state)
     }
 }
